+++ test/PowerShell.UnitTests/Extensions/EnumerableExtensions.cs
@@ -1,0 +1,28 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+namespace Microsoft.Store.PartnerCenter.PowerShell.UnitTests.Extensions
+{
+    using System;
+    using System.Collections.Generic;
+
+    /// <summary>
+    /// Extensions for enumerators.
+    /// </summary>
+    public static class EnumerableExtensions
+    {
+        /// <summary>
+        /// Perform an action on each element of a sequence.
+        /// </summary>
+        /// <typeparam name="T">Type of elements in the sequence.</typeparam>
+        /// <param name="sequence">The sequence.</param>
+        /// <param name="action">The action to perform.</param>
+        public static void ForEach<T>(this IEnumerable<T> sequence, Action<T> action)
+        {
+            foreach (T element in sequence)
+            {
+                action(element);
+            }
+        }
+    }
+}